Added React UI package
While there are quite some community-driven UI packages for React out there, many of them are either dated or unmaintained. This makes implementing accounts management on React rather cumbersome. It would be awesome to have an official package for all supported UI frameworks, as out-of-the-box user accounts are a major selling point for Meteor.